Add files via upload
Implement the function countryDetector, which takes a string as input representing the field country from the UK Company House and cleans the data and returns a normalized name of a country.

Often the field country in Company House was not clear with name of cities, counties or mixtures of cities and country. The function countryDetector is basically a data cleaner for countries.